% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMFM.R
\name{vMFM}
\alias{vMFM}
\title{Variable Selection for Von Mises–Fisher Mixtures}
\usage{
vMFM(x, k, max.iter = 100, tol = 1e-06, red.cutoff = Inf,
  red.max.consid = 0, red.method = "dist", noise.max.consid = 0,
  do.redundant = T, do.noise = T, use.movMF = F)
}
\arguments{
\item{x}{data.frame or matrix to perform clustering and variable selection}

\item{k}{Integer. The number of clusters.}

\item{max.iter}{Integer. The maximum number of iterations for each EM run.}

\item{tol}{Numeric. The tolerance value used to determine if convergence has been reached.}

\item{red.cutoff}{(optional) Numeric. the cutoff value to use for selecting redundant columns by distance.}

\item{red.max.consid}{(optional) integer. The number of redundant variable pairs to consider.}

\item{red.method}{"dist" (default), to select red.max.consid by minimum distance, "all" to select all variables, or "rand" to randomly select red.max.consid pairs}

\item{noise.max.consid}{(optional) integer. The number of noise variables to consider.}

\item{do.redundant}{logical. True (default) if redundant variables should be removed.}

\item{do.noise}{logical. True (default) if noise variables should be removed.}

\item{use.movMF}{logical. True if the movMF package should be used when possible. False (Default) to use built-in EM algorithm.}
}
\value{
A list object with the redundant variable pairs (if any), noise variables (if any), and final clustering.
}
\description{
Variable Selection for Von Mises–Fisher Mixtures
}
